+++ CHANGELOG.md
@@ -3,3 +3,6 @@
 **2.0.0-rc2**
 
   - added `data-parsley-validate-if-empty` field option (#489)
+  - fixed select multiple bug (#522)
+  - allowed checkbox, radio and select multiple inputs to have either a `name`
+    or an `id` to be binded (instead of just a name)